test(prefs): migrate DataStore tests from androidHostTest to commonTest
Migrate 3 core:prefs DataStore preference tests (FilterPrefsTest,
TakPrefsTest, NotificationPrefsTest — 12 test methods total) from
androidHostTest to commonTest for KMP compatibility.

Changes:
- Replace java.io.File with okio.Path/FileSystem for temp directories
- Replace PreferenceDataStoreFactory.create() with createWithPath()
- Replace JUnit4 TemporaryFolder with manual Okio temp dir + cleanup
- Replace org.junit.Assert with kotlin.test assertions
- Remove withHostTest block from core/prefs/build.gradle.kts
- Update architecture-review doc to reflect core:prefs test coverage
