+++ core/prefs/build.gradle.kts
@@ -24,7 +24,6 @@ kotlin {
     android {
         namespace = "org.meshtastic.core.prefs"
         androidResources.enable = false
-        withHostTest { isIncludeAndroidResources = true }
     }
 
     sourceSets {
